wifi: Add a Marvell variant to ConnectMBO
Marvell is a fullMAC driver, meaning we can't control what goes into the
association request IEs, and thus we can't include MBO IEs even if we
support MBO. This is more a pedantic part of the MBO spec than an actual
functionality check (we can still connect to the MBO AP fine), so get
rid of the IE check for Marvell devices.

BUG=b:184873062
TEST=run both variants on non-Marvell and Marvell devices and make sure
they pass/are skipped as expected.

Cq-Depend: chromium:3258348
Change-Id: Ibd5ba52a9228a949095d6412d9e1c4f9fd61e29c