tast: Add tests for mic mute
Map the string "micmute" to KEY_MICMUTE in key_mappings.go.

Add two test for mic mute toggle via the keyboard: one that uses the
debug accelerator (works on all devices) and one that uses the
dedicated keyboard key (works only on devices that have a MICMUTE
key).  In both tests, we mute/unmute the mic and examine the output of
cras_test_client to verify that the mic is in fact muted/unmuted.

BUG=chromium:1263278
TEST=Ran on both Volteer and MrBland EVT units

Change-Id: I871ba772873a332d8d806224d4fc57924fd5c8d5