wilco_ec: test_event uses ASCII string
Change the data written to test_event in sysfs to ASCII '0' instead of
raw byte '0'.

Between kernel 4.19 and 5.10, the function that parses the input
inside simple_attr_write changed from "simple_strtoll" to "ksstrtoull"
which has better error checking. ASCII "0" is correctly converted to
0ull, whereas the byte '0' results in a parsing error, causing
ksstrtoull to return -EINVAL. simple_strtoll also failed to parse byte
'0' but it didn't matter since the result value was default initialized
to 0 and no error checking ocurred. This passes on sarien with the 4.19
kernel and the 5.10 kernel.

Relevant change in kernel code:

https://source.chromium.org/chromiumos/_/chromium/chromiumos/third_party/kernel/+/488dac0c9237647e9b8f788b6a342595bfa40bda

BUG=b:203449710
TEST=tast run <ip> wilco.ECEvent

Change-Id: I9f6f21f95d285a0ea90384f7cfd14839ce8a0494