tast-tests: Add radioTapFCSValid filter to Dot11FCSValid
Frame check sequence(FCS) is either present or not in PCAP packets,
depending on devices and settings. Gale's PCAP capturing strips out
Dot11 FCS from all packets, so that RadioTap layer shows "FCS at end:
false". The wireless adapter checks the integrity of received packets
on-the-fly and reflect the status in RadioTap.Flags fields. If a packet
is broken, it's RadioTap.Flags.BadFCS should be set to true, otherwise
it's false.

If Dot11 FCS are not present in packet in a PCAP, gopacket will calculate
and append FCS by itself. In which case, the Dot11 FCS always matches
the checksum of actually captured packets, even though the integrity of
transmitted data is broken, due to gopacket's expected behavior.

Occasionally, due to sniffer reception error, it will capture corrupted
packets thus the PCAP would have bad FCS. In flaky ProbeReqFormat tests,
we found some packets encapsulating broken IEs still had valid Dot11
FCS. ie. One of Supported Rates is 0xdd, which is invalid, in packet
SN=2791. Previously Dot11FCSValid considered this packet good, but the
Dot11InformationElement decoder reported it was broken.

Check RadioTap FCS before Dot11 layer filter to tell if packets are
truly broken or not. The user API is unchanged.

BUG=b:185378075
TEST=tast run -var="router=${ROUTER}" -var="pcap=${PCAP}" ${DUT} \
	wifi.ProbeReqFormat, ran 100 times on volteer and trogdor
	each with all [PASS]
TEST=./src/platform/tast/fast_build.sh -T

Change-Id: Iab78292fe6448ea8419a8d9c220ecd873d74507c