Tast: Increase ARCVM video encode test timeout.
On most platforms the current timeout of 4min is enough for the ARCVM
video encoder tests. But on some lower spec boards such as certain
o ctopus variants this timeout is not enough resulting in test
failures.

On octopus the arc.VideoEncodeAccel.h264_1080p_i420_vm test takes about
8min, so this CL increases the timeout  to 12min to leave a 50% margin
for fluctuations.

BUG=b:205060753
TEST=arc.VideoEncodeAccelPerf.h264_*_i420_vm on octopus

Change-Id: I3576bb9965dc37daddd98ea5337478f73e0693b1