kernel: Make pstore console ramoops test less flaky
Sometimes the message that this test writes to the kernel log isn't
present in the console-ramoops file because the ramoops file is full of
other kernel messages or errors/warnings that happened between when
/dev/kmsg was written and reboot actually rebooted. Let's just test to
confirm that the file exists and has data. The file is only populated by
the kernel if the ramoops is valid from the previous kernel, so we don't
really need to confirm that the file contents have anything we wrote
from the previous kernel.

In fact, we can skip the reboot path entirely if the file is already
present as it means console ramoops worked for _this_ kernel boot,
making this test much faster to run in the general case. If the file
isn't there for some reason, for example if userspace deleted the
ramoops file from pstore, then we'll run reboot and check for the file.
This should reduce the flakyness of this test and we should be able to
promote it to critical.

BUG=b:189058499
TEST=tast run $DUT kernel.PstoreConsoleRamoops

Change-Id: I7a3128019ab74f5f97d2c06abab86059fac85864