network: VPNConnect: wait for charon exited during cleanup
Currently, shill does not have a Disconnecting state for a vpn service,
and thus when the service is in the Disconnected (Idle) state, the VPN
driver may still be doing some clean up. For an L2TP/IPsec service, that
means the charon process might still be running, and then the next
connect attempt will fail if there is another charon process running.
This case is rare in the real world but very likely to happen in the
tast tests. This patch adds the logic to wait until the charon process
exits for each test run, to avoid the flakiness caused by it.

BUG=None
TEST=Ran `network.VPNConncect.l2tp_ipsec*` multiple times, all green.

Change-Id: If201cbf0565d31d151897aeffadb304f29dc0d8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3261197
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Alexander Hartl <alexanderhartl@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>