arcappcompat: Update Tast test for Disney
BUG=b:205035708
TEST=tast run <eve> arcappcompat.Disney.clamshell_mode

It is observed that on some devices after the login page,
Disney has a new dialog window for handling the never button
followed by a new window for the profile icon page. Added a
code to handle it. Also Ui ID for my stuff and logout got changed.
Updated it.

Change-Id: I99b1f8e21c2099687fb24070015d837b5e2ed611
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3263256
Tested-by: Meenakshi Thiyagarajan <mthiyagarajan@google.com>
Reviewed-by: Renuga Nanjappan <rnanjappan@chromium.org>
Commit-Queue: Meenakshi Thiyagarajan <mthiyagarajan@google.com>