Update the import path of aosp/perfetto proto go package
Context CL:
https://android-review.git.corp.google.com/c/platform/external/perfetto/+/1881638
https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3262643

We updated the go_package path in perfetto protos in perfetto repo. We also have
a pending CL to update the ebuild of dev-go/perfetto-protos in
chromiumos repo. Once we uprev the ebuild, the import paths in the
related go files must be adjusted acorrdingly at the same time.

BUG=b:204473963
TEST=Local test

Change-Id: Iddd76f7e8867653436f3c328571a0e3d08301842