tast-tests: Update SMB private directory
The filemanager.SMB test is currently failing with unable to locate the
private directory.

/var/lib/samba/private/msg.sock: No such file or directory

The top level directory is not created on deployment due to the files
not being required for the parts that ChromeOS utilises. To get past
this we can override the private dir in smb.conf.

BUG=chromium:1156844
TEST=tast -verbose run <DUT-IP> filemanager.SMB

Change-Id: I4e9a0d650cf704b4a7e3594bd999cf23365dc44b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3267586
Tested-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>