tast-tests: note how RTD2142s may spuriously fail
fwupd_detect_rtd2142 may fail on Puff devices when the device has an old
firmware installed on the MST, which software cannot differentiate from
an actual test failure. To permit users to better understand these kinds
of failures if they find a device that has old firmware, add a message
to the failure message pointing to more information.

BUG=b:173742142
TEST=builds

Change-Id: I13ae64f02da62873c36d295bb047337eb5049f59