crossdevice: Refactor BeTo e2e tests.
This CL does the following

1. Makes common package small and clear.
Trying to figure out where code was located was a nightmare so now
the common package *only* has stuff that is *directly* used by local
and remote now. It is one function a bunch of constants.
Also everytime we access one of the common parts ive used
nearbycommon.blah or crossdevicecommon.blah to make it extra explicit.

2. I've moved nearbysnippet and nearbytestutil back to local :D
The stuff that couldn't be moved I took out of them and put in common.
99% of nearbytestutil was local stuff anyway.

3. crossdevice is now the source of truth for shared Android setup code.
nearbyshare setup calls into there as much as possible.
nearbysetup is deleted. crossdevice did almost all of it.

4. Nested Attributes
So crossdevice is the basic attributes. Then nearbyshare adds in their own extra
fields. Maybe over-optimizing but it works ok.
The Android snippet wrappers call into a central function to get the
basic Android attributes. The nearby adds in the few nearby fields.

5. GoogleAccountUtil always comes from crossdevice.
Made nearbyshares data file a symlink.

BUG=b:200705945
TEST=tast run -build lab-dut '("group:nearby-share")'
TEST=tast run -build lab-dut '("group:cross-device")'
TEST=tast run -build lab-dut '("group:nearby-share-remote")'
Exempt-From-Owner-Approval: keep losing +2

Change-Id: I371caa4271c387efe0ba07dc175422ce66e0f66d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3269031
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>