tast-tests: Restart ui for shill tests
Shill clones the root mount namespace at the time of respawn and
propagates all mounts as private mounts. As a result, if shill is
restarted when a cryptohome is mounted, it will keep the cryptohome
mounts open perpetually, since unmount events will not be propagated
inside this namespace.

In practice, this affects the unmount/removal for dm-crypt cryptohomes
where the underlying dm-crypt device cannot be removed until the mount
exists (which is the case till shill is restarted next). Instead,
restart ui before resetting shill: this ensures that the next test
does not need to worry about the mounts in shill's mount namespace.

BUG=b:172344853
TEST=network.DiagFailDNSResolution passes when run twice.

Change-Id: Iaccde3c9efbf5e059b3f2fcff1a0c14d014a8fba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3267499
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>