arcappcompat: use intents to start hover test
Adjust the mouse hover test to start based on broadcasted intents. Prior
to this change, the hover test needed a 'Start Test' button to be
clicked in order to ensure the mouse wouldn't trigger a hover event
before the test began. Using the mouse to start the test could be
flakey, and made the hover test have essentially two assertions:
clicking, and hovering. This change makes the test state handled by
intents, which are much more reliable.

BUG=b:204490158
TEST=push apk && tast run ${HOST} arc.StandardizedMouseHover*

Change-Id: I5e266e3d86274f941d3f5b32d44a62463656de7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3273214
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Tested-by: David Welling <davidwelling@google.com>
Commit-Queue: David Welling <davidwelling@google.com>