[Lacros] Move user data clean up code to fixture.
This CL moves the user data clean up code for launching lacros from UI
to the fixture code. This is because we are about to land the keep
lacros alive with app service, which will make the lacros started at
session login, and clean up the user data before launching lacros from
UI will break the test, because the user data will not be created due to
the launch.

BUG=chromium:1268743
TEST=test the modified tests still pass

Change-Id: I37fa83b227610b449a1b041527951426f0680d3a