tast-tests: Update expectation for Chrome PIP maximum size
This CL fixes failures of ui.ChromePIPEnergyAndPower.big_lacros by
updating the expectation for the maximum size of the PIP window. The old
expectation only ever worked for the one board on which I manually
verified. The new expectation will be relatively permissive to maximize
code maintainability. For simplicity, the same expectation will be used
in ui.ChromePIPEnergyAndPower.big.

BUG=chromium:1269802
TEST=tast run <DUT> ui.ChromePIPEnergyAndPower.big*

Change-Id: I878731e9155850b6a43ba0cc45528a70cf8fd529