arcappcompat: If display power is off fail the test, Alive func to check UiAutomator Server status
It is observed that one of the reasons for the “Failed to start
the activity of an app” issue is due to “Failed to ensure the
display is on”. So replaced log message with fatal. Also observed
that “Failed to get the current app package” error is due to the
UiAutomator server disconnection. Added d.Alive to make sure the
UiAutomator server is up. Also replaced error logs for “Failed to
get window info” with fatal to get the exact screenshot for the failure.

BUG=b:183135675, b:200214797
TEST=tast run <helios> arcappcompat.Starz.vm_clamshell_mode
         tast run <helios> arcappcompat.Starz.vm_tablet_mode
         tast run <kevin> arcappcompat.Starz.clamshell_mode
         tast run <kevin> arcappcompat.Starz.tablet_mode

Change-Id: I8b3b1a955cb021c3a5116daa138a1d3ca1737c98
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3276502
Commit-Queue: Meenakshi Thiyagarajan <mthiyagarajan@google.com>
Tested-by: Meenakshi Thiyagarajan <mthiyagarajan@google.com>
Reviewed-by: Renuga Nanjappan <rnanjappan@chromium.org>
(cherry picked from commit abd33c9f92ad18af79a4a778ca75a5334dc502c1)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3313843