arcappcompat: Fix for failed to start the activity of an app issue
It is observed that there are two cases when Failed to start the
activity of an app issue occurs, in the first case if the test cases
are failing due to “"Failed to get current app package and Failed to
get window info" or Failed initializing UI Automator: UI Automator
server did not come up:" then the following testcases are failing
with “Failed to start the activity of an app” error due to “Failed
to ensure the display is on”. In the second case, if a test is failing
due to “ Failed to start the activity of an app"  then following tests
are failing with “Failed to start the activity of an app” error due to
“Failed to ensure the display is on”. Both Case 1 and case 2 are not
either app issues or platform issues of the device. It is happening only
when running the tast tests and it isn’t reproducible manually.
So to avoid the flaky test failures, if the above errors are observed,
the arcappcompat tests will skip the test instead of failing flaky.
Also ran and tested this CL on all apps under arcappcompat suite.
It is working fine as intended.

BUG=None
TEST=tast run <helios> arcappcompat.ABCKids.vm_clamshell_mode
         tast run <helios> arcappcompat.ABCKids.vm_tablet_mode
         tast run <kevin> arcappcompat.ABCKids.clamshell_mode
         tast run <kevin> arcappcompat.ABCKids.tablet_mode

Change-Id: I8b3b1a955cb021c3a5116daa138a1d3ca1737c98