wifi: add wifi softAP capability test
This change adds `iw list` check for SoftAP capabilities. This test checks
if DUT supports softAP, max number of associated STAs supported.

This change also seperates the HE capabilities for STA and AP in
accordance with iw and nl80211. So when DUT supports HE in STA mode,
the test checks if DUT also supports HE in SoftAP mode.

BUG=b:205914971

TEST:tast run <dut-ip> wifi.SAPCaps (trogdor)
TEST:tast run <dut-ip> wifi.Caps wifi.Caps80211ac wifi.HeCaps (trogdor, hatch)

Cq-Depend: chromium:3282756
Change-Id: I341b0bc9c493f2cbc171abe234489385b558ede5