Suppress Lacros experimental infobar
This CL adds --test-type when launching Lacros to suppress
the experimental infobar ("Experimental (alpha-quality) browser! Some
features missing or incomplete...") and possible others as well that
could cause noise when benchmarking.

Infobars are not displayed when launching ash-chrome and, in order to
ensure consistency, now they won't be show when launching Lacros as
well.

BUG=None
TEST=tast run device lacros.Basic

Change-Id: I940db3c18decff35530c1a4df8cb8127613cb5ad