Suppress Lacros experimental infobar
This CL adds --test-type when launching Lacros and LacrosUI fixtures to
suppress the experimental infobar ("Experimental (alpha-quality)
browser! Some features missing or incomplete...") that could cause noise
when benchmarking.

Infobars are not displayed when launching ash-chrome and in order to
ensure consistency now they won't be shown either when launching Lacros.

BUG=None
TEST=tast run $device lacros.Basic
TEST=tast run $device lacros.ShelfLaunch

Change-Id: I940db3c18decff35530c1a4df8cb8127613cb5ad