arcappcompat:Update Tast test for Adobe Acrobat Reader
Adobe Acrobat Reader is failing in launch.
On some devices, it is observed that with
the older version of code, to select and
click on the gmail account, tab key and
enter key are pressed before the gmail
account page appears. Added a code to wait
for the gmail account page to appear and
then perform tab key and enter key to
select and click on gmail account will
fix the flaky failure.

BUG=b:205774065
TEST=tast run <eve> arcappcompat.AdobeAcrobatReader

Change-Id: I695152837b9f67ddb13b9a0b87f495a9d508ccbf