tast-tests: Send JS to foreground instead of background page
The Files app background page defers its start until the foreground page
is loaded for the first time. This means the fixture can't access the
page unless the Files app has already begun. Further to that when the
Files app migrates to the SWA there will be no background page. Move
this logic to the foreground page preemptively.

BUG=chromium:1156844
TEST=tast run <DUT-IP> filemanager.SMB

Change-Id: I5912398c1ce545a031b304304dcc6adb7df2f730