wifi.PTK: Bump timeouts
We're occasionally getting timeouts in the ping runner, and yet the
stdout logging shows that it finished 149 or 150 of 150 ICMP packets.
Presumably there's a little extra work in ping command teardown that's
just missing the timeout, so let's add another buffer there too.

While we're at it: this test performs on the order of 1-2 minutes of
activity, with potentially another XX seconds of setup and teardown,
depending on the fixture behavior and external hardware (APs). Let's put
a real number in here, and make it conservative. Tastboard shows that
*successful* runs occasionally reach >300 seconds.

BUG=b:205800361
TEST=run it

Change-Id: I2f3c358a2a5d0a692b6b7932bc3179411c091af2