faft: Add var firmware.skipFlashUSB
Add new var firmware.skipFlashUSB, which defaults to false. Setting it
prevents the test from downloading the OS image and flashing to the
USB stick. This is useful when running a test on a skylab machine. The
OS image on the usb stick is probably close enough, and it takes forever
to flash when running remotely.

BUG=b:206684999
TEST=tast run --var=servo=${SERVO?} --var firmware.skipFlashUSB=true \
$D firmware.BootMode.dev_to_rec

Change-Id: I78ae571c1bff1a8c3686921f1a4adbf35a636173
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3290113
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Andrew Luo <aluo@chromium.org>