typec: Add veyron boards to basic test blocklist
veyron_tiger and veyron_fievel don't have Type C ports. The Basic test
gets scheduled on these boards because the EC doesn't provide a response
to the "ectool inventory" command. Ordinarily, this is WAI, since it
leads to a test failure which is an indicator of an issue in the EC, but
in this case, it's a false negative. So, blocklist these two boards from
this test.

BUG=b:206672864
TEST=None

Change-Id: Idac7aa1dc8faf1c7b75f95c3e1095964b2980fac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3287291
Tested-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>