wifi: Keep scanning until DUT is connected to the AP
Sometimes DUT cannot connect to the AP due to service not found, which
leads to flakiness in test. Move scan request to the polling loop
instead of trying for a fixed number of attempts. DUT will keep scanning
until it is connected to the AP or when polling timeouts.

BUG=b:203034912
TEST=pass 10 iterations of `tast run -var=router=<Router IP> <DUT IP> wifi.PolicyBasic`

Change-Id: Iac8ec81b0ba5516c81e2e6438881e1bbc3777cb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3288848
Tested-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Jun Yu <junyuu@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Commit-Queue: Jintao Lin <jintaolin@chromium.org>