camera: Make sure test app is closed when fails to pass CCAUIIntent
Currently there is a common issue that when a sub test of CCAUIIntent
fails, it is very likely that all the following sub tests will fail
since the state is not clean up properly. There are two findings:
1. Pending app window is not dropped on Chrome side
2. Test app is not closed if it fails in cca.Init().

This CL tries to close test app no matter the sub test passes/fails to
solve (2).

BUG=b:204019671
TEST=tast run [DUT] camera.CCAUIIntent

Change-Id: Iccc3f8579948724bd84e3f7ad0306b36ce203dcd