arc.Print: Wait for printer ready notification
arc.Print.* has occasional flakes where the printer name can't be found
in the dropdown.  It isn't clear why these happen, but it's possible
that the printer isn't quite ready when the test app is launched.  Since
Chrome shows a notification when printer setup is complete, wait for
this notification before starting the app.  This should cause no
slowdown in the common case, but lets the test wait for the printer when
setup takes longer than usual.

BUG=b:159325335
TEST=300+ iterations of tast run arc.Print.* on several boards

Change-Id: I9cfe1f04d6276c96bc169a93047e59c3928e81db