faft: Add ssh based GBB functions
Add new GetGBBFlags and ClearAndSetGBBFlags funcs which uses ssh instead
of a gRPC service. gRPC services are a problem when booting from USB
drives, as the TAST gRPC service binary is not present and has to be
copied to the machine. Most tests that boot from USB don't actually need
any services except the ones required to get booted into the right mode.

BUG=b:202758040
TEST=firmware.BootMode.dev_to_rec

Change-Id: Ib22e68c1d0ebbde41eb3b0fe501e45e3e9fe2d40