faft: Don't use any services in fixture
Convert the 2 uses of services into ssh calls in the firmware.fixture.*
and dependent code.

gRPC services are a problem when booting from USB drives, as the TAST
gRPC service binary is not present and has to be copied to the machine.
Most tests that boot from USB don't actually need any services.

BUG=b:202758040
TEST=firmware.BootMode* firmware.Fixture*

Change-Id: If41f43e1252096e5a85527094f3902a52231c2f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3290115
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Andrew Luo <aluo@chromium.org>