faft: Only copy tast files when explicitly needed
Change ModeSwitcher to only copy tast files if the CopyTastFiles option
is passed.

Change firmware.fixture to pass CopyTastFiles in the cases where it is
unlikely any copying would actually be necessary.

Rename the USB fixtures to make it clear that services won't be
available.

Change test firmware.Fixture to not use the BiosService.

BUG=b:202758040
TEST=firmware.BootMode* firmware.Fixture* firmware.BootedDeviceReporter*
firmware.Eventlog*

Change-Id: Ibf229080ca5801426f53b85a581c0051191f66cb