usbprinter: Add kernel module fixture
All tests that use the `virtual-usb-printer` must load the required
kernel modules. This is currently expressed as two repeated boilerplate
function calls in the bodies of each test. This CL converts the
boilerplate into a fixture, eliminating the need to mention the modules
except as a `Fixture:` parameter in initializing the test.

This CL is purely ergonomic; we do not currently suspect that module
churn (loading and unloading the modules once per test) causes any
particular flakiness or consumes resources unnecessarily.

BUG=b:206732481
TEST=tast run <dut> '("dep:virtual_usb_printer")'

Disallow-Recycled-Builds: test-failures
Change-Id: Ib0cf141fbd75f24aa59e273ef79adb513e691942
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3290680
Commit-Queue: Kalvin Lee <kdlee@chromium.org>
Tested-by: Kalvin Lee <kdlee@chromium.org>
Reviewed-by: Jun Sheng <shengjun@chromium.org>