video: Reshuffle AV1 video.ChromeStackDecoding test cases
video.PlatformDecoding.av1* has 3x2 test cases:
- av1_common: Verifies the common files using MD5SUM criteria.
- av1_film_grain: Verifies the "film grain" using MD5SUM
- av1_ssim: Verifies all the above using SSIM.

(x2 is because this three groups are run for 8 and for
10 bits).

The second entry av1_film_grain is not run on volteer, but
that's the only platform supporting AV1 decoding, so this is
effectively not run anywhere. That doesn't matter because
the test vectors are run as part of the third entry av1_ssim.

Since SSIM is a weaker criteria than MD5SUM, this CL leaves
the cases as:
- av1_coommon verifies all-non-grain-film using MD5SUM.
- av1_film_grain using SSIM.

BUG=b:205354844
TEST=tast run <IP> video.ChromeStackDecoding.av1* on volteer

Change-Id: I29c348c915e56d307164531168bbb0660ffa9e49