camera: Add logs to gather information of screenshot issue
We found that sometimes the screenshot mechanism does not work so that
we cannot understand the current situation of CCA from test results well.

This CL adds some logs to help us understand if it is not triggered or
it is an issue in screenshot functionality.

BUG=b:204527195
TEST=tast run [DUT] camera.CCAUIStress.real

Change-Id: I7663a5e5ce43a97bc6ec57b327bdb2a825dc7e7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3290777
Tested-by: Wei Lee <wtlee@chromium.org>
Auto-Submit: Wei Lee <wtlee@chromium.org>
Reviewed-by: Inker Kuo <inker@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Wei Lee <wtlee@chromium.org>