tast-tests: use protoreflect to avoid unexported fields
As github.com/golang/protobuf uprev to v1.5.2, the current usage of
reflect breaks since proto fields are default to unexported.  This CL
uses the suggested protoreflect library to solve this.

This CL also extend the timeout for the report generation because some
probe configs contains more and more possible comps to probe.

BUG=b:206876222
TEST=tast run "${DUT_IP}" platform.CrosHardwareVerifier

Change-Id: I7ce151f32a5535c5faddcb7dddaf1f302c756e8b