tast-tests: Stabilize launcher.BubbleSmoke
The test flakily fails ~1% of the time. This seems to happen more often
on tablet-first devices, but I cannot reproduce it. We theorize the UI
might be unstable if the test has to transition from tablet to
clamshell mode at test start. Add the same ui.WaitForLocation() fix
used in launcher.Smoke and other tests.

Also fix a typo in an error message.

BUG=chromium:1254420
TEST=tast run <dut> launcher.BubbleSmoke

Change-Id: I4af3936114b84640a30467a16a88103e32f179b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3291861
Tested-by: James Cook <jamescook@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>