health: add ProbeWifiInfo ,health: add MonitorThunderboltEvent, health:ProbeBusInfo tests
Check that we can probe cros_healthd for WIFI info.
Monitors the Thunderbolt event detected proper or not after plug/unplug Thunderbolt devices.
Check that we can probe cros_healthd for bus info and Thunderbolt data.
In this CL we have added the supported board names.
And We are disabling data access protection for Thunderbolt devices.

BUG=b:198745197
TEST=Ran on DUT locally

Change-Id: I226e69ec70de0a1e0b7e9e45d99dd4d642e6a166