factory.Finalize: Disable from all runs
factory.Finalize just promoted with running in an environment which is
more close to real world, but found there is an incompatible networking
change in factory wiping tool that will break connections and the rest
test can not run. Disable it until the wiping tool is compatible with
the tast lab settings.

BUG=b:202931696
TEST=tast run <DUT> factory.Finalize

Change-Id: I0aa3edd6a427bef7370eeb1c1c20cc3609f342b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3291218
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Reviewed-by: Meng-Huan Yu <menghuan@chromium.org>
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>
Tested-by: Sung-Chi Li <lschyi@chromium.org>