cups: Remove unused argument
The `DevInfo` struct is unused in `cups.RestartPrintingSystem()`. This
CL removes the unused argument and trivially reorders the calls in tests
that use `cups.RestartPrintingSystem()` directly.

The reordering is done to emphasize that at the changed call sites, the
`DevInfo` struct can be loaded _after_ the printing system is restarted
without changing any important behavior in the test. This in turn is
important because it unlocks a potential future CL in which
`usbprinter.LoadPrinterIDs()` can be folded into the rest of the
boilerplate that starts the `virtual-usb-printer`.

BUG=b:206732481
TEST=tast run <dut> '("dep:virtual_usb_printer")'

Change-Id: I3600a7c695ed587c934137df84c0c91fac57de44