Tast: Add WifiNotMarvell dependency to IWScan
Add the hardware dependency WifiNotMarvell to the Tast test wifi.IWScan,
so that the test will be skipped on devices with Marvell WiFi. These
devices have known issues with this test, and for reasons described in
$BUG, we will not be pursuing fixes.

This change also removes the software dependency no_elm_hana_3_18, since
it is made redundant by the new hardware dependency.

BUG=b:172218718
TEST=tast run $DUT wifi.IWScan
skips correctly on elm, runs on non-Marvell devices

Change-Id: I8e9881c16cbdf06e429b77105a17ba39d4e1379e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3293962
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>