Lacros: Disable what's new feature in Lacros tests
The What's-New feature automatically redirects the browser to a WebUI
page to display the new feature if this is first time the user opens
the browser or the user has upgrade Chrome to a different milestone.

This CL disables the feature in testing to make the test expectations
more predirectable, and thus make the tests more stable.

BUG=chromium:1271436
TEST=Ran Lacros critical Tast tests

Change-Id: Icfa4032fb46deeb41e78c508cebf489d15608b61
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3295310
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Hyung Tae Kim <hyungtaekim@chromium.org>
Tested-by: Yuke Liao <liaoyuke@chromium.org>
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>