productivity cuj: Refine clean up logic
Reconnecting the chrome target in the cleanup prompt should consider whether the service tab is already open.
Otherwise, MemoryInfoTracker will continue scanning until the case timeout is exceeded.
Change to use the defer function instead of the CleanUp function to close all browser tabs so that the cleanup can be activated even if the test cases fail.

BUG=b:208340318
TEST=tast run <vars> <dut> ui.MicrosoftOfficeWebCUJ.*
TEST=tast run <dut> ui.GoogleDocsWebCUJ.*

Change-Id: I79ab02e34f32e44cca2671078a8375a178d37c65