Harden some parts of the Android setup.
Nesting fixtures means a single step failing in the Android fixture can
bring down an entire run.

This CL adds retries around the parts of the Android setup that has
failed in the last month: adbsetup and clearing logcat logs.

I've resisted adding retries everywhere so let's go with
this for now.

BUG=b:207152986
TEST=nearbyshare.PhoneToCrosHighVis.dataonline_noone_txt30mb

Change-Id: I65c6f13c12711169cc130d00aa3458e4b41651c3