arcappcompat: Adjust window bounds to ensure pointer coordinates
The chrome.autotestPrivate.mouseMove method does not consistently
position the mouse after switching a window from 'Fullscreen' to
'Normal'. This causes flakiness in the mouse, and trackpad related tests
of the standardized test suite. To account for this, the window bounds
are now adjusted after converting to 'Normal', in order to allow the
mouse to be positioned correctly. This was verified by running the test
on repeat multiple times.

BUG=b:194900373
TEST=./repeat-test.sh -s -r 100 -- $DUT arc.StandardizedMouse*

Change-Id: Ia170368cd07a7d6b6a26de048ffe7d08d631074b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3300799
Tested-by: David Welling <davidwelling@google.com>
Commit-Queue: David Welling <davidwelling@google.com>
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Reviewed-by: Matt Stark <msta@google.com>
Reviewed-by: Seewai Fu <seewaifu@google.com>