faft: Remove optimization in dev->usb dev reboot
When booting to usb dev mode, you have to go to dev mode first, then
enable usb booting, and then reboot again.  The code skips this process
if the DUT is already in dev mode. However this doesn't work if the
DUT was in dev mode only because of GBB flags.

Remove this optimization, so that DevGBB -> USBDev transitions work.

BUG=b:207698262
TEST=firmware.Fixture.dev_usb

Change-Id: Id4eaa34fc479feba43110e445ec3e3225ceb750f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3301081
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Kshitij Shah <tij@google.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>