camera: Reset test bridge between sub tests in CCAUIIntent
We will reset test bridge between sub tests for other camera tests but
currently this mechanism is missing on CCAUIIntent since it uses its own
way to start/stop the app.

Having this mechanism, when a sub test fails, it will be less likely to
affect the following sub tests.

BUG=b:207738292
TEST=tast run [DUT] camera.CCAUIIntent

Change-Id: If3deece6062d2c1a1baafcfe4a8fcf4d912ac4b0