tast-tests: Update filesapp to provide alternate SWA launch path
The Files app is undergoing a migration from a Chrome app to a System
Web App. As part of this migration the filesapp uiauto library requires
uplift to work with both variants. Whilst rolling out the Finch
configuration this might see test DUTs use either the Chrome app or SWA
unless we're explicit.

Update the filesapp library to include a path to launch (and close) the
Files app. This ensures all call sites that connect to Files app after
it has launched also call the correct variant.

BUG=b:207576612
TEST=./fast_build.sh -T

Change-Id: I22d344b17463d5135d7f3020b617129e3e0d878e