tast-tests: Update filesapp to work with SWA variant
The Files app is undergoing a migration from a Chrome app to a System
Web App. As part of this migration the filesapp uiauto library requires
uplift to work with both variants. Whilst rolling out the Finch
configuration this might see test DUTs use one or the other so for
critical tests be explicit about the variant it uses until the SWA
variant has completely launched.

BUG=b:207576612
TEST=tast run <DUT-IP> ("group:mainline" && !informational && "dep:chrome")

Change-Id: I22d344b17463d5135d7f3020b617129e3e0d878e