Tast: Fix metadata for the tast.arc.VideoDecodeAccelVDVDA.vp9_2 test.
The ARCVM VideoDecodeAccelVDVDA.vp9_2 test is currently failing as the
bit_depth field required by the video_decode_accelerator_tests is
missing from the test data json file. The vp9_2 json test data is
out-of-sync with the Chrome data, this CL makes sure we now use the
latest version that has the bit_depth field.

BUG=b:207813282
TEST=tast.arc.VideoDecodeAccelVDVDA.vp9_2

Change-Id: I6be0893a6622d974237100d7b4cbad79c3108976