crostini: Debian OS Upgrade UI alert.
Automated test to check that the UI alerts the user that the
crostini linux container is no longer supported and should be
upgraded to the latest version. This is achieved by creating a
mock /etc/os-release file that refers to stretch, then
restarting the crostini container to trigger the ui alert.

Test params are generated using the minimal set, which for now
is Buster only. When Debian OS next releases a new stable, this
test will need to be updated.

On cleanup the original os-release file that is symlinked
to /usr/lib/os-release is restored.

BUG=b:205193011
TEST=Ran on hatch DUT and ony hatch lab DUTs.

Change-Id: I12a51f41ec3d81d1e6224c16876f5b28f06ed7b9