crostini: Debian OS Upgrade UI alert.
Automated test to check that the UI alerts the user that the
crostini linux container isno longer supported and should be
upgraded to the latest version. This is achieved by creating a
mock /etc/os-release file that refers to stretch, then
restarting the crostini container to trigger the ui alert.

On cleanup the original os-release file that is symlinked
to /usr/lib/os-release is restored.

BUG=b:205193011
TEST=Ran on hatch DUT and ony hatch lab DUTs.

Change-Id: I12a51f41ec3d81d1e6224c16876f5b28f06ed7b9