dlp: Fix JSON generation for arrays
Mark any array fields as "omitempty" such that a missing value does not
produce a null value and thus an ill-formed policy.

In particular, this fixes a validation issue in the
DataLeakPreventionRulesList policy tests (but potentially in others
too).  Based on this, a follow-up CL will make use of
policyutil.ServeAndVerify where possible.

BUG=chromium:1274514
TEST=tast run dlp.DataLeakPreventionRulesList*

Change-Id: I326ab13acb24743ec3d2175fd7819871f10f5b33
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3306381
Tested-by: Georg Neis <neis@chromium.org>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Commit-Queue: Georg Neis <neis@chromium.org>