dlp: Fix JSON generation for DataLeakPreventionRulesList policy
Mark the Components field of
DataLeakPreventionRulesListValueDestinations as "omitempty" such that
a missing Components value does not produce a null value and thus an
ill-formed policy. This seems to have been a validation-only issue,
i.e., there should be no change in behavior.

Based on this, make use of policyutil.ServeAndVerify where possible.

BUG=chromium:1274514
TEST=tast run dlp.DataLeakPreventionRulesList*

Change-Id: I326ab13acb24743ec3d2175fd7819871f10f5b33