audio: Add loopback lifecycle tests
This CL add tests for multiple lifecycle combinations of the loopback
device. For example the loopback starts without a playback stream, a
playback stream stops when the loopback device is capturing audio, etc.

The correct behavior of the loopback device is to continue to provide
audio samples when there are no playback streams.

BUG=b:199262343
TEST=None

Change-Id: I42f29e7fc7a3b11d89947e77bcda8152c91be43b